Tier B Phase 3d-ii-b: relocate app-tools registry + tool-input streaming
Second step of the renderer/cluster relocation. app-tools-registry (SEP-1865
App-Provided Tools registry) and useToolInputStreaming move into
@mcpjam/widget-react verbatim (they were already @/-free after 3b — only
ext-apps/client/react/zustand + the ./widget-host contract).

- Inspector app-tools-registry.ts + useToolInputStreaming.ts become re-export
  shims; every import site (use-chat-session, useToolExecution, modal/renderer,
  tests) is unchanged. The registry's zustand store singleton lives in the
  package and is shared via the re-export.
- Package gains zustand as a dep (externalized in tsup); index re-exports both
  modules.

Verified: package guard/typecheck/test/build (dist 28.6 KB, dts external refs
only); inspector typecheck:client + 219 tests (incl. use-chat-session +
useToolExecution exercising the registry via shim); root typecheck. No behavior
change.
