+++ widget-react/src/index.ts
@@ -24,6 +24,9 @@ export {
   readToolResultMeta,
   readToolResultServerId,
 } from "./tool-result-utils";
+// App-provided tools registry (SEP-1865) + tool-input streaming (3d-ii-b).
+export * from "./app-tools-registry";
+export * from "./useToolInputStreaming";
 export type {
   WidgetHost,
   // primitives
